# [system-probe] Use sync.Pool for http aggregation objects