--- a/pkg/network/encoding/format_test.go
+++ b/pkg/network/encoding/format_test.go
@@ -183,7 +183,10 @@ func BenchmarkFormatHTTPStats(b *testing.B) {
 
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		FormatHTTPStats(in)
+		result := FormatHTTPStats(in)
+		for _, a := range result {
+			releaseHTTPAggregations(a)
+		}
 	}
 }